+++ lua/haskell-tools/lsp.lua
@@ -13,6 +13,7 @@ local uv = vim.uv
 --- * `:HlsStart` - Start the LSP client.
 --- * `:HlsStop` - Stop the LSP client.
 --- * `:HlsRestart` - Restart the LSP client.
+--- * `:HlsEvalAll` - Evaluate all code snippets in comments.
 ---@brief ]]
 
 ---To minimise the risk of this occurring, we attempt to shut down hls cleanly before exiting neovim.
@@ -237,6 +238,13 @@ local commands = {
     'HlsRestart',
     function()
       HlsTools.restart()
+    end,
+    {},
+  },
+  {
+    'HlsEvalAll',
+    function()
+      HlsTools.buf_eval_all()
     end,
     {},
   },